+++ udata/__init__.py
@@ -4,5 +4,5 @@
 udata
 '''
 
-__version__ = '6.2.1.dev'
+__version__ = '7.0.1.dev'
 __description__ = 'Open data portal'