+++ udata/core/dataset/apiv2.py
@@ -234,7 +234,7 @@ class DatasetExtrasAPI(API):
         # then update the extras with the remaining payload
         dataset.extras.update(data)
         try:
-            dataset.save()
+            dataset.save(signal_kwargs={'ignores': ['post_save']})
         except mongoengine.errors.ValidationError as e:
             apiv2.abort(400, e.message)
         return dataset.extras
@@ -254,7 +254,7 @@ class DatasetExtrasAPI(API):
                 del dataset.extras[key]
         except KeyError:
             apiv2.abort(404, 'Key not found in existing extras')
-        dataset.save()
+        dataset.save(signal_kwargs={'ignores': ['post_save']})
         return dataset.extras, 204
 
 
@@ -352,7 +352,7 @@ class ResourceExtrasAPI(ResourceMixin, API):
             data.pop(key)
         # then update the extras with the remaining payload
         resource.extras.update(data)
-        resource.save()
+        resource.save(signal_kwargs={'ignores': ['post_save']})
         return resource.extras
 
     @apiv2.secure
@@ -371,5 +371,5 @@ class ResourceExtrasAPI(ResourceMixin, API):
                 del resource.extras[key]
         except KeyError:
             apiv2.abort(404, 'Key not found in existing extras')
-        resource.save()
+        resource.save(signal_kwargs={'ignores': ['post_save']})
         return resource.extras, 204