+++ udata/core/dataset/models.py
@@ -309,7 +309,7 @@ class ResourceMixin(object):
     @property
     def last_modified(self):
         if self.harvest and self.harvest.modified_at and to_naive_datetime(self.harvest.modified_at) < datetime.utcnow():
-            return max([self.last_modified_internal, to_naive_datetime(self.harvest.modified_at)])
+            return to_naive_datetime(self.harvest.modified_at)
         if self.filetype == 'remote' and self.extras.get('analysis:last-modified-at'):
             return to_naive_datetime(self.extras.get('analysis:last-modified-at'))
         return self.last_modified_internal
@@ -632,7 +632,7 @@ class Dataset(WithMetrics, BadgeMixin, db.Owned, db.Document):
     def last_modified(self):
         if (self.harvest and self.harvest.modified_at and
                 to_naive_datetime(self.harvest.modified_at) < datetime.utcnow()):
-            return max([self.last_modified_internal, to_naive_datetime(self.harvest.modified_at)])
+            return to_naive_datetime(self.harvest.modified_at)
         return self.last_modified_internal
 
     @property